fix: prevent integer overflow in sequence number validation
- Change last_sequence_number_ from int to int64_t to handle edge case
  at ~2B messages without overflow risk
- Add explicit cast in comparison to prevent signed overflow
- Remove unnecessary std::move in connection_manager erase (no-op)
- Add clarifying comment to queue_message [[nodiscard]]

+++ src/server/websocket_session.hpp
@@ -3,6 +3,7 @@
 #include "boost_wrapper.hpp"
 #include <atomic>
 #include <chrono>
+#include <cstdint>
 #include <memory>
 #include <mutex>
 #include <deque>
@@ -85,10 +86,11 @@ class websocket_session final
   std::atomic<bool> should_close_{false};
   void do_close();
 
-  [[nodiscard]] bool queue_message(std::string&& message);
+  [[nodiscard]] bool queue_message(std::string&& message);  // Returns false if queue full
 
   // Sequence number tracking for replay attack prevention
-  std::atomic<int> last_sequence_number_{-1};
+  // Using int64_t to prevent overflow edge case at INT_MAX sequences
+  std::atomic<int64_t> last_sequence_number_{-1};
 
   // Rate limiting for DoS prevention
   // Sliding window using timestamps of recent messages
